requirements.txt
openai~=0.27.6
python-telegram-bot
python-dotenv~=1.0.0